+++ application/models/Fleets.php
@@ -7,7 +7,9 @@
  *
  * @author chris
  */
-class Fleet extends CI_Model
+
+class Fleets extends CI_Model
+
 {
 
 	var $data = array(